+++ tests/conf.py
@@ -10,7 +10,8 @@ import time
 
 import niquests
 
-from . import compatibility_issues
+from caldav import compatibility_hints
+from caldav.davclient import CONNKEYS
 from caldav.davclient import DAVClient
 
 ####################################
@@ -140,7 +141,7 @@ if test_radicale:
             "username": "user1",
             "password": "",
             "backwards_compatibility_url": url + "user1",
-            "incompatibilities": compatibility_issues.radicale,
+            "incompatibilities": compatibility_hints.radicale,
             "setup": setup_radicale,
             "teardown": teardown_radicale,
         }
@@ -225,35 +226,16 @@ if test_xandikos:
             "name": "LocalXandikos",
             "url": url,
             "backwards_compatibility_url": url + "sometestuser",
-            "incompatibilities": compatibility_issues.xandikos,
+            "incompatibilities": compatibility_hints.xandikos,
             "setup": setup_xandikos,
             "teardown": teardown_xandikos,
         }
     )
 
+
 ###################################################################
 # Convenience - get a DAVClient object from the caldav_servers list
 ###################################################################
-## TODO: this is already declared in davclient.DAVClient.__init__(...)
-## TODO: is it possible to reuse the declaration here instead of
-## duplicating the list?
-## TODO: If not, it's needed to look through and ensure the list is uptodate
-CONNKEYS = set(
-    (
-        "url",
-        "proxy",
-        "username",
-        "password",
-        "timeout",
-        "headers",
-        "huge_tree",
-        "ssl_verify_cert",
-        "ssl_cert",
-        "auth",
-    )
-)
-
-
 def client(
     idx=None, name=None, setup=lambda conn: None, teardown=lambda conn: None, **kwargs
 ):
@@ -266,8 +248,8 @@ def client(
         return client(**caldav_servers[idx])
     elif name is not None and no_args and caldav_servers:
         for s in caldav_servers:
-            if caldav_servers["name"] == s:
-                return s
+            if s["name"] == name:
+                return client(**s)
         return None
     elif no_args:
         return None